Customized Readme
Took out now irrelevant info from readme. It should be noted that the readme had instruction from Udacity which provided the starter code for this project.

From Udacity:
The starter project has some HTML and CSS styling to display a static version of the Memory Game project. You'll need to convert this project from a static project to an interactive one. This will require modifying the HTML and CSS files, but primarily the JavaScript file.